Tests al 100% mi pez

+++ s3_escarabajos-web/src/main/java/co/edu/uniandes/csw/escarabajos/dtos/ReclamoDTO.java
@@ -15,7 +15,7 @@ import java.util.List;
  */
 public class ReclamoDTO 
 {
-    private Integer id;
+    private Long id;
     private List<String> pics;
     private String mensaje;
     private String razon;
@@ -52,7 +52,7 @@ public class ReclamoDTO
     {
         razon = raz;
     }
-     public Integer getId()
+     public Long getId()
     {
         return id;
     }
